requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
Babel==2.13.1
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
emoji==2.12.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
magic-filter==1.0.12
multidict==6.0.5
nest-asyncio==1.6.0
openai==0.28.0
psycopg2-binary==2.9.9
pyaes==1.6.1
pybabel==0.0.0.dev0
pydantic==2.8.2
pydantic_core==2.20.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.34
TgCrypto==1.2.5
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.11
